+++ packages/core/phase2/exceptions/checkResultsMatchingPncDisposalsExceptions.ts
@@ -1,9 +1,9 @@
 import type { AnnotatedHearingOutcome, Offence, Result } from "../../types/AnnotatedHearingOutcome"
-import isMatchToPncDisposal from "../lib/areAllResultsOnPnc/isMatchToPncAdjudicationAndDisposals/isMatchToPncDisposal"
+import isMatchToPncDisposal from "../lib/isMatchToPncDisposal"
 import type { PncDisposal } from "../../types/PncQueryResult"
 import isRecordableResult from "../lib/isRecordableResult"
-import createPncAdjudicationFromAho from "../lib/areAllResultsOnPnc/isMatchToPncAdjudicationAndDisposals/createPncAdjudicationFromAho"
-import isMatchToPncAdjudication from "../lib/areAllResultsOnPnc/isMatchToPncAdjudicationAndDisposals/isMatchToPncAdjudication"
+import createPncAdjudicationFromAho from "../lib/createPncAdjudicationFromAho"
+import isMatchToPncAdjudication from "../lib/isMatchToPncAdjudication"
 import findPncCourtCase from "../lib/findPncCourtCase"
 
 type CheckExceptionFn = (result: Result, offenceIndex: number, resultIndex: number) => void
