+++ packages/core/phase2/lib/areAllResultsOnPnc/isMatchToPncAdjudicationAndDisposals/index.ts
@@ -1,1 +1,1 @@
-export { default as isMatchToPncAdjudicationAndDisposals } from "./isMatchToPncAdjudicationAndDisposals"
+export { default as isMatchToPncAdjudicationAndDisposals } from "../../isMatchToPncAdjudicationAndDisposals"
