+++ packages/core/phase2/lib/areResultsMatchingAPncDisposal.ts
@@ -1,6 +1,6 @@
-import type { Offence } from "../../../../types/AnnotatedHearingOutcome"
-import type { PncDisposal } from "../../../../types/PncQueryResult"
-import isRecordableResult from "../../isRecordableResult"
+import type { Offence } from "../../types/AnnotatedHearingOutcome"
+import type { PncDisposal } from "../../types/PncQueryResult"
+import isRecordableResult from "./isRecordableResult"
 import isMatchToPncDisposal from "./isMatchToPncDisposal"
 
 const areResultsMatchingAPncDisposal = (offence: Offence, disposals: PncDisposal[]): boolean =>
