+++ packages/core/phase2/lib/isMatchToPncAdjudicationAndDisposals.ts
@@ -1,8 +1,8 @@
-import type { AnnotatedHearingOutcome, Offence } from "../../../../types/AnnotatedHearingOutcome"
+import type { AnnotatedHearingOutcome, Offence } from "../../types/AnnotatedHearingOutcome"
 import areResultsMatchAPncDisposal from "./areResultsMatchingAPncDisposal"
 import createPncAdjudicationFromAho from "./createPncAdjudicationFromAho"
 import isMatchToPncAdjudication from "./isMatchToPncAdjudication"
-import findPncCourtCase from "../../findPncCourtCase"
+import findPncCourtCase from "./findPncCourtCase"
 
 const isMatchToPncAdjudicationAndDisposals = (aho: AnnotatedHearingOutcome, offence: Offence): boolean => {
   const offenceReasonSequence = offence.CriminalProsecutionReference?.OffenceReasonSequence ?? undefined
