+++ packages/core/phase2/lib/isMatchToPncDisposal.test.ts
@@ -1,5 +1,5 @@
-import type { Result } from "../../../../types/AnnotatedHearingOutcome"
-import type { PncDisposal } from "../../../../types/PncQueryResult"
+import type { Result } from "../../types/AnnotatedHearingOutcome"
+import type { PncDisposal } from "../../types/PncQueryResult"
 import isMatchToPncDisposal from "./isMatchToPncDisposal"
 
 describe("isMatchToPncDisposal", () => {
