+++ packages/core/phase2/lib/isMatchToPncDisposal.ts
@@ -1,6 +1,6 @@
-import type { Result } from "../../../../types/AnnotatedHearingOutcome"
-import type { PncDisposal } from "../../../../types/PncQueryResult"
-import { createPncDisposalFromResult } from "../../createPncDisposalFromResult"
+import type { Result } from "../../types/AnnotatedHearingOutcome"
+import type { PncDisposal } from "../../types/PncQueryResult"
+import { createPncDisposalFromResult } from "./createPncDisposalFromResult"
 
 const isMatchToPncDisposal = (pncDisposals: PncDisposal[], result: Result): boolean =>
   createPncDisposalFromResult(result).every((ahoDisposal) =>
